% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{print_program}
\alias{print_program}
\title{Print a specific program}
\usage{
print_program(program_name)
}
\arguments{
\item{program_name}{Name of the program to print}
}
\description{
Print a specific program
}
