% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\docType{data}
\name{program_code}
\alias{program_code}
\title{Program code collection}
\format{
An object of class \code{list} of length 10.
}
\usage{
program_code
}
\description{
A list containing various R program examples
}
\keyword{datasets}
